+++ packages/shared/components/src/Toast/index.ts
@@ -1,0 +1,3 @@
+/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
+export { Toast, type ToastProps } from './toast';
+export { ToastProvider } from './toast.provider';
